pyproject.toml
[tool.poetry]
name = "oops"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
alembic = "^1.13.1"
psycopg2 = "^2.9.9"
pydantic = "^2.7.4"
python-dotenv = "^1.0.1"
flask-pydantic = "^0.12.0"
flask-openapi3 = {extras=["swagger"], version = "^3.1.3"}
pydantic-settings = "^2.3.3"


[build-system]
requires = ["poetry<=1.8.3"]
build-backend = "poetry.core.masonry.api"
